MovieLens-Dataset/requirements.txt
# 基础依赖
pandas>=1.5.0
numpy>=1.23.0
tqdm>=4.64.0
requests>=2.28.0

# 深度学习和NLP
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0

# 向量检索
faiss-cpu>=1.7.4
# 如果有GPU，使用: faiss-gpu>=1.7.4

# 关键词检索（可选）
rank-bm25>=0.2.2

# 数据处理
scikit-learn>=1.2.0

# API和Web（可选）
flask>=2.3.0
fastapi>=0.95.0
uvicorn>=0.22.0